Renamed Toggle Inspector command and moved it to View menu
