+++ src/main/java/sc/iview/commands/MenuWeights.java
@@ -56,7 +56,6 @@ public final class MenuWeights {
     public static final double EDIT_TOGGLE_FLOOR = 50;
     public static final double EDIT_DELETE_OBJECT = 100;
     public static final double EDIT_PROPERTIES = 200;
-    public static final double EDIT_SCENE = 300;
 
     public static final double PROCESS_ISOSURFACE = 0;
     public static final double PROCESS_CONVEX_HULL = 1;
@@ -72,6 +71,8 @@ public final class MenuWeights {
     public static final double VIEW_RESET_CAMERA_ROTATION = 202;
     public static final double VIEW_RESET_CAMERA_POSITION = 203;
     public static final double VIEW_SAVE_CAMERA_CONFIGURATION = 204;
+    public static final double VIEW_TOGGLE_INSPECTOR = 302;
+
 
     public static final double DEMO_LINES = 0;
     public static final double DEMO_MESH = 1;
